refactor(shared): move cross-feature utilities to @/shared/utils
Move utilities used by 3+ features to shared/utils to eliminate duplication:

- playoffWeight: moved from transactions/utils to shared/utils/calculations
- formatDelta: removed duplicate from matchups, now uses shared version
- median: consolidated to use shared version from stats utils

Updated all imports in:
- features/transactions/utils/facts.ts
- features/matchups/utils/swing-analysis.ts
- features/stats/components/TrendsView/utils.ts
- app/league/transactions/page.tsx
- app/stats/utils/computeTransactionGradesForStatsHub.ts
- features/matchups/components/SwingPointsDisplay.tsx

Added comprehensive README.md documenting shared utils usage.

Closes #15
